+++ CHANGELOG.md
@@ -32,6 +32,8 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix shader error when clipping flags are set without clip objects present
 - Breaking: Add `Volume.instances` support and a `VolumeInstances` transform to dynamically assign it
   - This change is breaking because all volume objects require the `instances` field now.
+- Fix wrong group count calculation on geometry update (#1562)
+- Fix wrong instance index in `calcMeshColorSmoothing`
 
 ## [v4.18.0] - 2025-06-08
 - MolViewSpec extension: